GetAll, GetOne, Add, Mod and Del DBA profile completed
To do so, many changes and additions has been done.
For example:
As ModDBAProfile uses an interactive texts asking for the inputs,
addAutoResponse function has been added in QTelnetInterface class
As DBAProfile speeds configuration is no fixed but up to DBA type and
the values must be multiple of 64, custom Spin boxes has been added and
especial behaviour in GUI coded.